Load fonts using WOFF2 format.
WOFF2 provides provides a 23.4% reduction in file size for the Fira Code
Regular and Bold fonts.

According to CanIUse, 95.52% of browsers support WOFF2 at this point,
when compared to 98.28% for WOFF.

Aka, pretty much every major browser across desktop and mobile support
WOFF2, and we provide a fallback to WOFF for those few browsers that
don't.